requirements.txt
Flask
psycopg[binary]
python-dotenv
coverage